requirements.txt
requests~=2.28.1
torch~=1.11.0
Pillow==9.3.0
typing~=3.7.4.3
torchvision~=0.12.0
opencv-python~=4.5.5.64
numpy~=1.22.4
loguru~=0.6.0
uvicorn~=0.17.6
fastapi~=0.78.0
starlette~=0.19.1
pydantic~=1.9.1
click~=8.1.3
tqdm~=4.64.0
setuptools~=62.3.2
aiofiles~=0.8.0
python-multipart~=0.0.5
timer-cm==1.1
runpod==0.9.1
